posting bawang

+++ database/migrations/2021_10_06_051835_create_products_table.php
@@ -16,7 +16,7 @@ class CreateProductsTable extends Migration
         Schema::create('products', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->tinyInteger('stock');
+            $table->integer('stock');
             $table->integer('price');
             $table->date('harvest_time');
             $table->text('description');
